+++ src/cache/event.rs
@@ -3,64 +3,8 @@ use std::num::NonZeroU16;
 
 use extract_map::entry::Entry;
 
-use super::{BaseGuildChannel, Cache, CacheUpdate, GenericChannelId, GuildThread};
-use crate::all::{
-    CountDetails,
-    MessageReaction,
-    ReactionAddEvent,
-    ReactionRemoveAllEvent,
-    ReactionRemoveEmojiEvent,
-    ReactionRemoveEvent,
-};
-use crate::internal::prelude::*;
-use crate::model::channel::{ChannelInfoChannel, GuildChannel, Message};
-use crate::model::event::{
-    ChannelCreateEvent,
-    ChannelDeleteEvent,
-    ChannelInfoEvent,
-    ChannelPinsUpdateEvent,
-    ChannelUpdateEvent,
-    GuildCreateEvent,
-    GuildDeleteEvent,
-    GuildEmojisUpdateEvent,
-    GuildMemberAddEvent,
-    GuildMemberRemoveEvent,
-    GuildMemberUpdateEvent,
-    GuildMembersChunkEvent,
-    GuildRoleCreateEvent,
-    GuildRoleDeleteEvent,
-    GuildRoleUpdateEvent,
-    GuildScheduledEventCreateEvent,
-    GuildScheduledEventDeleteEvent,
-    GuildScheduledEventUpdateEvent,
-    GuildStickersUpdateEvent,
-    GuildUpdateEvent,
-    MessageCreateEvent,
-    MessageUpdateEvent,
-    PresenceUpdateEvent,
-    ReadyEvent,
-    ThreadCreateEvent,
-    ThreadDeleteEvent,
-    ThreadListSyncEvent,
-    ThreadUpdateEvent,
-    UserUpdateEvent,
-    VoiceChannelStartTimeUpdateEvent,
-    VoiceChannelStatusUpdateEvent,
-    VoiceStateUpdateEvent,
-};
-use crate::model::gateway::Presence;
-use crate::model::guild::{
-    Guild,
-    GuildMemberFlags,
-    Member,
-    MemberCount,
-    MemberGeneratedFlags,
-    Role,
-    ScheduledEvent,
-};
-use crate::model::id::GuildId;
-use crate::model::user::{Collectibles, CurrentUser, OnlineStatus};
-use crate::model::voice::VoiceState;
+use super::{Cache, CacheUpdate};
+use crate::model::prelude::*;
 
 impl CacheUpdate for ChannelCreateEvent {
     type Output = GuildChannel;
