+++ src/http/mod.rs
@@ -26,8 +26,6 @@ mod ratelimiting;
 mod request;
 mod routing;
 
-use std::sync::Arc;
-
 use reqwest::Method;
 pub use reqwest::StatusCode;
 
@@ -37,67 +35,7 @@ pub use self::multipart::*;
 pub use self::ratelimiting::*;
 pub use self::request::*;
 pub use self::routing::*;
-#[cfg(feature = "cache")]
-use crate::cache::Cache;
 use crate::model::id::*;
-
-/// This trait will be required by functions that need [`Http`] and can optionally use a [`Cache`]
-/// to potentially avoid REST-requests.
-///
-/// If the `cache` feature is enabled, but an implementing type does not have access to a cache,
-/// the [`CacheHttp::cache`] method will simply return `None`.
-pub trait CacheHttp: Send + Sync {
-    fn http(&self) -> &Http;
-
-    #[cfg(feature = "cache")]
-    #[must_use]
-    fn cache(&self) -> Option<&Arc<Cache>> {
-        None
-    }
-}
-
-impl<T> CacheHttp for &T
-where
-    T: CacheHttp,
-{
-    fn http(&self) -> &Http {
-        (*self).http()
-    }
-    #[cfg(feature = "cache")]
-    fn cache(&self) -> Option<&Arc<Cache>> {
-        (*self).cache()
-    }
-}
-
-impl<T> CacheHttp for Arc<T>
-where
-    T: CacheHttp,
-{
-    fn http(&self) -> &Http {
-        (**self).http()
-    }
-    #[cfg(feature = "cache")]
-    fn cache(&self) -> Option<&Arc<Cache>> {
-        (**self).cache()
-    }
-}
-
-#[cfg(feature = "cache")]
-impl CacheHttp for (Option<&Arc<Cache>>, &Http) {
-    fn cache(&self) -> Option<&Arc<Cache>> {
-        self.0
-    }
-
-    fn http(&self) -> &Http {
-        self.1
-    }
-}
-
-impl CacheHttp for Http {
-    fn http(&self) -> &Http {
-        self
-    }
-}
 
 /// An method used for ratelimiting special routes.
 ///
